refactor: Building main window
